CS now sets up signal handlers sooner
Signal handlers are now registered (and signals are processed) while
other resources are still being allocated.
